+++ Project/gearworks/src/lib/engine/Engine.cpp
@@ -6,15 +6,18 @@
 // If they require other libraries that have not yet been loaded, initialize the variables in the Initialize function
 Rectangle *rect;
 
-Engine::Engine(unsigned int *_mainShaderProgram, VertexArrayObject *_vao)
-	: mainShaderProgram(_mainShaderProgram), vao(_vao) {
+VertexArrayObject rectVAO;
+
+Engine::Engine(unsigned int *_mainShaderProgram)
+	: mainShaderProgram(_mainShaderProgram) {
 }
 Engine::~Engine() {}
 
 // Initialization logic here
 void Engine::Initialize() {
 	// Rectangle
-	rect = new Rectangle(vao, mainShaderProgram, glm::vec2(0.0f, 0.0f), glm::vec2(400.0f, 296.0f), "resources/textures/logo.png");
+	rectVAO.Initialize();
+	rect = new Rectangle(&rectVAO, mainShaderProgram, glm::vec2(0.0f, 0.0f), glm::vec2(400.0f, 296.0f), "resources/textures/logo.png");
 }
 
 // Updating variables etc
@@ -24,6 +27,7 @@ void Engine::Update() {
 
 // Rendering shapes etc
 void Engine::Render() {
+	rectVAO.Bind();
 	rect->Render();
 }
 
@@ -31,4 +35,6 @@ void Engine::Render() {
 void Engine::Clean() {
 	delete rect;
 	rect = NULL;
+
+	rectVAO.Unbind();
 }